+++ backend/hardware/devicelist.py
@@ -2,7 +2,7 @@ import hardware.stirring as stirrer
 import hardware.reagentdispenser as rd
 import hardware.temperaturecontroller as tc
 import hardware.thermometer as thermometer
-import hardware.gpiochip as gpiochip
+import hardware.gpiochip.core as gpiochip
 from config import microlabConfig as config
 import yaml
 from os.path import exists
